requirements.txt
dj-database-url==0.5.0
Django==3.1.6
django-cors-headers==3.5.0
django-heroku==0.3.1
django-leaflet @ git+https://github.com/makinacorpus/django-leaflet.git@844887affe607d3f115920c862f8ea2b45e19ed8
django-otp==1.0.0
django-tinymce==3.0.2
django-tinymce4-lite==1.8.0
djangorestframework==3.11.2
jsmin==2.2.2
mysqlclient==1.4.6
Pillow==8.1.1
psycopg2==2.8.5
python-dotenv==0.14.0
pytz==2019.3
six==1.13.0
sqlparse==0.3.1
whitenoise==5.1.0
django-filter==2.4.0
django-extensions
